laser-suite/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "laser-suite"
version = "0.1.0"
description = "Deterministic geodetic adjustment and ALTA/NSPS compliance pipeline"
readme = "README.md"
requires-python = ">=3.11"
authors = [{name = "LASER Suite"}]
dependencies = [
  "numpy>=1.26",
  "PyYAML>=6.0",
  "shapely>=2.0",
]

[project.scripts]
laser-suite = "laser_suite.cli:main"

[tool.setuptools]
package-dir = {"" = "python"}

[tool.setuptools.packages.find]
where = ["python"]

[tool.pytest.ini_options]
pythonpath = ["python"]
testpaths = ["python/tests"]
